Refactor exception to inherit main "WellNus" exception
