+++ src/main/java/seedu/duke/WellNus.java
@@ -1,6 +1,8 @@
 
 package seedu.duke;
 
+import seedu.duke.atomichabit.feature.AtomicHabitManager;
+
 import java.util.Scanner;
 
 public class WellNus {
