--- src/main/java/seedu/duke/atomichabit/exception/AtomicHabitException.java
@@ -1,7 +1,0 @@
-package seedu.duke.atomichabit.exception;
-
-public class AtomicHabitException extends Exception {
-    public AtomicHabitException(String errorMessage) {
-        super(errorMessage);
-    }
-}
