+++ src/main/java/seedu/duke/exception/AtomicHabitException.java
@@ -1,0 +1,9 @@
+package seedu.duke.exception;
+
+import seedu.duke.exception.WellNusException;
+
+public class AtomicHabitException extends WellNusException {
+    public AtomicHabitException(String errorMessage) {
+        super(errorMessage);
+    }
+}
